Add custom encoder && decoder stub

+++ gen.sh
@@ -2,17 +2,19 @@
 
 # Simple bash script to generate shellcode from assembly code
 # > bash gen.sh file.asm
+# > nasm -felf64 execve_sh.asm -o a.out
+# > objdump -M intel -D a.out
 
-# > nasm -felf64 execve_sh.nasm -o a.out
+nasm -felf64 $1 -o a.out && objdump -D -M intel | grep -e " 00" && 
 
-nasm -felf64 $1 -o a.out && objdump -D -M intel | grep -e " 00" && echo -e "\nShellcode [\033[91mnull byte!\033[0m]:\n" ||
+echo -e "\n\033[91m[-]\033[0m Shellcode |\033[41m\033[97m Null Byte! \033[0m|:\n" ||
 
-echo -e "Shellcode [\033[92mOK\033[0m]:\n"
+echo -e "\n\033[92m[+]\033[0m Shellcode |\033[42m\033[97m PASS \033[0m|:\n"
 
-for i in $(objdump -D -M intel | grep "^ " | cut -f2)
+for i in $(objdump -D | grep "^ " | cut -f2)
 do
-	echo -n "\x${i}"
+	[ $i == 00 ] && echo -ne "\033[93m\\\x${i}\033[0m" || echo -n "\x${i}"
 done
 
-rm a.out
+[ -f a.out ] && rm a.out
 echo
